+++ src/Console/DefaultCommandsTrait.php
@@ -3,6 +3,7 @@
 namespace Solital\Core\Console;
 
 use Solital\Core\Console\MessageTrait;
+use Solital\Core\Kernel\Application;
 
 trait DefaultCommandsTrait
 {
@@ -135,6 +136,6 @@ trait DefaultCommandsTrait
 
         echo $about . $version . $date . PHP_EOL . PHP_EOL;
         $this->line("PHP Version (" . PHP_VERSION . ")")->print()->break();
-        $this->line("By Solital Framework. Access http://solitalframework.com/")->print()->break();
+        $this->line("By Solital Framework. Access " . Application::SITE_DOC_DOMAIN)->print()->break();
     }
 }